third_party/WebKit/LayoutTests/platform/win/svg/animations/beginEndAnimationElement-use-counter-expected.txt

This is a testharness.js-based test.
PASS endElement is use counted 
Harness: the test ran to completion.
